IGNITE-25169 Make BinarySchemaImpl package-private
