+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/utils/PlatformUtils.java
@@ -1229,7 +1229,7 @@ public class PlatformUtils {
                 for (int j = 0; j < fieldCnt; j++)
                     fieldIds.add(reader.readInt());
 
-                schemas.add(new BinarySchema(id, fieldIds));
+                schemas.add(BinaryUtils.binarySchema(id, fieldIds));
             }
         }
