+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -288,6 +288,7 @@ org.apache.ignite.internal.binary.BinaryObjectOffheapImpl
 org.apache.ignite.internal.binary.BinaryReaderExImpl$Flag
 org.apache.ignite.internal.binary.BinarySchema
 org.apache.ignite.internal.binary.BinarySchema$Confirmation
+org.apache.ignite.internal.binary.BinarySchemaImpl
 org.apache.ignite.internal.binary.BinaryTreeMap
 org.apache.ignite.internal.binary.BinaryWriteMode
 org.apache.ignite.internal.binary.builder.BinaryLazyMap$1$1$1
